resume-builder-backend/requirements.txt
# requirements.txt
anyio==4.3.0 # or similar, dependency of fastapi/starlette
bcrypt==4.1.3 # or similar, dependency of passlib
certifi==2024.2.2 # or similar, dependency of motor/pymongo
dnspython==2.6.1 # or similar, dependency of motor/pymongo
email-validator==2.1.1 # or similar, dependency of pydantic for EmailStr
fastapi==0.111.0 # or similar
h11==0.14.0 # or similar, dependency of uvicorn
idna==3.7 # or similar, dependency of anyio/httpx
motor==3.4.0 # or similar
passlib==1.7.4 # or similar
pydantic==2.7.1 # or similar
pydantic-core==2.18.2 # or similar, dependency of pydantic
pydantic-settings==2.2.1 # or similar
python-dotenv==1.0.1 # or similar
python-jose[cryptography]==3.3.0 # or similar
pymongo==4.7.2 # or similar, dependency of motor
sniffio==1.3.1 # or similar, dependency of anyio/httpx
starlette==0.37.2 # or similar, dependency of fastapi
typing_extensions==4.11.0 # or similar, dependency of pydantic/fastapi
uvicorn==0.29.0 # or similar
# 可能还有其他间接依赖
